+++ src/bitop.rs
@@ -203,3 +203,225 @@ impl SmolBitSet {
         shortening_bitop_fn_body!(self, rhs, op);
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::bst_slice::BstSlice;
+
+    #[cfg(not(feature = "std"))]
+    use extern_alloc::vec::Vec;
+
+    /// Helper extension trait to simplify some tests.
+    trait AndNot {
+        fn and_not(self, rhs: Self) -> Self;
+    }
+
+    impl AndNot for BitSliceType {
+        fn and_not(self, rhs: Self) -> Self {
+            self & !rhs
+        }
+    }
+
+    impl AndNot for usize {
+        fn and_not(self, rhs: Self) -> Self {
+            self & !rhs
+        }
+    }
+
+    fn to_bst_vec(sbs: &SmolBitSet) -> Vec<BitSliceType> {
+        BstSlice::new(sbs).slice().to_vec()
+    }
+
+    fn zero_pad_to(mut vec: Vec<BitSliceType>, len: usize) -> Vec<BitSliceType> {
+        while vec.len() < len {
+            vec.push(0);
+        }
+        vec
+    }
+
+    mod inline {
+        use super::*;
+
+        macro_rules! test_inline_binops {
+                ($($name:ident, $a:expr, $b:expr),+) => {$(
+                    #[test]
+                    fn $name() {
+                        let a = SmolBitSet::from($a);
+                        let b = SmolBitSet::from($b);
+                        assert!(a.is_inline());
+                        assert!(b.is_inline());
+
+                        let res = a.$name(&b);
+                        assert!(res.is_inline());
+
+                        let d = unsafe { res.get_inline_data_unchecked() };
+                        assert_eq!(d, ($a as usize).$name($b as usize));
+                    }
+                )*}
+            }
+
+        test_inline_binops! {
+            bitor, 0xF0F0u16, 0x0F0Fu16,
+            bitand, 0xFEFEu16, 0xAFFEu16,
+            bitxor, 0x3A52u16, 0xAAE8u16,
+            and_not, 0xFEFEu16, 0xAFFEu16
+        }
+    }
+
+    mod slice {
+        use super::*;
+
+        macro_rules! test_slice_binops {
+                ($($name:ident, $a:expr, $b:expr),+) => {$(
+                    #[test]
+                    fn $name() {
+                        let a = SmolBitSet::from($a);
+                        let b = SmolBitSet::from($b);
+                        assert_eq!(a.len(), 2);
+                        assert_eq!(b.len(), 2);
+
+                        let res = a.$name(&b);
+                        assert_eq!(res.len(), 2);
+                        assert_eq!(
+                            res.as_slice(),
+                            [
+                                (($a as BitSliceType).$name($b as BitSliceType)),
+                                ((($a >> 32) as BitSliceType).$name(($b >> 32) as BitSliceType))
+                            ]
+                        );
+                    }
+                )*}
+            }
+
+        test_slice_binops! {
+            bitor, 0xC0FF_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            bitand, 0xC0FF_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            bitxor, 0xC0FF_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            and_not, 0xC0FF_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64
+        }
+    }
+
+    mod mixed {
+        use super::*;
+
+        macro_rules! test_mixed_binops {
+                ($($name:ident, $a:expr, $b:expr),+) => {$(
+                    #[test]
+                    fn $name() {
+                        let a = SmolBitSet::from($a);
+                        let b = SmolBitSet::from($b);
+                        assert!(a.is_inline());
+                        assert_eq!(b.len(), 2);
+
+                        let res1 = a.clone().$name(&b);
+                        assert_eq!(
+                            to_bst_vec(&res1),
+                            [
+                                (($a as BitSliceType).$name($b as BitSliceType)),
+                                ((($a >> 32) as BitSliceType).$name(($b >> 32) as BitSliceType))
+                            ]
+                        );
+
+                        let res2 = b.$name(&a);
+                        assert_eq!(res2, res1);
+                    }
+                )*}
+            }
+
+        test_mixed_binops! {
+            bitor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            bitand, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            bitxor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64
+        }
+
+        // `a.and_not(b) != b.and_not(a)`
+        #[test]
+        fn and_not() {
+            const A: u64 = 0x0ABC_EE00_1337_BEEFu64;
+            const B: u64 = 0xF00D_BEEF_0420_BEEFu64;
+
+            let a = SmolBitSet::from(A);
+            let b = SmolBitSet::from(B);
+            assert!(a.is_inline());
+            assert_eq!(b.len(), 2);
+
+            let res1 = a.and_not(&b);
+            assert_eq!(
+                to_bst_vec(&res1),
+                [
+                    ((A as BitSliceType).and_not(B as BitSliceType)),
+                    (((A >> 32) as BitSliceType).and_not((B >> 32) as BitSliceType))
+                ]
+            );
+        }
+    }
+
+    mod rhs_smaller {
+        use super::*;
+
+        macro_rules! test_rhs_smaller_binops {
+                ($($name:ident, $a:expr, $b:expr),+) => {$(
+                    #[test]
+                    fn $name() {
+                        let mut lhs = SmolBitSet::from($a);
+                        let rhs = SmolBitSet::from($b);
+                        lhs <<= 32u8;
+                        let lhs_len = lhs.len();
+                        assert!(lhs_len > rhs.len());
+
+                        let res = lhs.$name(&rhs);
+                        assert_eq!(
+                            to_bst_vec(&res),
+                            [
+                                ((0 as BitSliceType).$name($b as BitSliceType)),
+                                ((($a) as BitSliceType).$name(($b >> 32) as BitSliceType)),
+                                ((($a >> 32) as BitSliceType).$name(0 as BitSliceType))
+                            ]
+                        );
+                    }
+                )*};
+            }
+
+        test_rhs_smaller_binops! {
+            bitor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            bitand, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            bitxor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            and_not, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64
+        }
+    }
+
+    mod rhs_larger {
+        use super::*;
+
+        macro_rules! test_rhs_larger_binops {
+                ($($name:ident, $a:expr, $b:expr),+) => {$(
+                    #[test]
+                    fn $name() {
+                        let lhs = SmolBitSet::from($a);
+                        let mut rhs = SmolBitSet::from($b);
+                        rhs <<= 32u8;
+                        let rhs_len = rhs.len();
+                        assert!(rhs_len > lhs.len());
+
+                        let res = lhs.$name(&rhs);
+                        assert_eq!(
+                            zero_pad_to(to_bst_vec(&res), 3),
+                            [
+                                (($a as BitSliceType).$name(0 as BitSliceType)),
+                                ((($a >> 32) as BitSliceType).$name($b as BitSliceType)),
+                                ((0 as BitSliceType).$name(($b >> 32) as BitSliceType))
+                            ]
+                        );
+                    }
+                )*};
+            }
+
+        test_rhs_larger_binops! {
+            bitor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            bitand, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            bitxor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
+            and_not, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64
+        }
+    }
+}
