+++ src/from.rs
@@ -84,3 +84,52 @@ impl FromStr for SmolBitSet {
         Ok(sbs)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    macro_rules! test_from {
+            ($($name:ident, $val:expr),+) => {$(
+                #[test]
+                fn $name() {
+                    let t = SmolBitSet::from($val);
+                    assert!(t.is_inline());
+
+                    let d = unsafe { t.get_inline_data_unchecked() };
+                    assert_eq!(d, $val as usize);
+                }
+            )*}
+        }
+
+    test_from! {
+        u8, 0b1110_1010u8,
+        u8_max, u8::MAX,
+        u16, 0xBEAFu16,
+        u16_max, u16::MAX,
+        u32, 0xF0BA_B0BBu32,
+        u32_max, u32::MAX,
+        u64, 0x3550_1337_0000_F00Fu64
+    }
+
+    #[test]
+    fn u64_hb_64() {
+        let t = SmolBitSet::from(0xC5C5_BEEF_0000_1234u64);
+        assert!(!t.is_inline());
+        assert_eq!(t.len(), 2);
+
+        let d = t.as_slice();
+        assert_eq!(d.len(), 2);
+        assert_eq!(d, [0x0000_1234, 0xC5C5_BEEF]);
+    }
+
+    #[test]
+    fn u64_max() {
+        let t = SmolBitSet::from(u64::MAX);
+        assert_eq!(t.len(), 2);
+
+        let d = t.as_slice();
+        assert_eq!(d.len(), 2);
+        assert_eq!(d, &[u32::MAX; 2]);
+    }
+}
