+++ src/lib.rs
@@ -35,7 +35,6 @@
 extern crate alloc as extern_alloc;
 #[cfg(not(feature = "std"))]
 use {
-    core::hash,
     core::slice,
     extern_alloc::alloc::{self, Layout, handle_alloc_error},
 };
@@ -43,7 +42,6 @@ use {
 #[cfg(feature = "std")]
 use {
     std::alloc::{self, Layout, handle_alloc_error},
-    std::hash,
     std::slice,
 };
 
@@ -66,10 +64,14 @@ mod bst_slice;
 mod cmp;
 mod fmt;
 mod from;
+mod hash;
 mod shifts;
 
 #[cfg(feature = "serde")]
 mod serde;
+
+#[cfg(feature = "typesize")]
+mod typesize;
 
 type BitSliceType = u32;
 
@@ -505,76 +507,14 @@ fn slice_layout(len: usize) -> Layout {
     layout
 }
 
-impl hash::Hash for SmolBitSet {
-    fn hash<H: hash::Hasher>(&self, state: &mut H) {
-        if self.is_inline() {
-            unsafe { self.get_inline_data_unchecked() }.hash(state);
-            return;
-        }
-
-        let hb = self.highest_set_bit();
-        let data = unsafe { self.as_slice_unchecked() };
-        for d in data.iter().take(hb.div_ceil(BST_BITS)) {
-            d.hash(state);
-        }
-    }
-}
-
-#[cfg(feature = "typesize")]
-impl typesize::TypeSize for SmolBitSet {
-    fn extra_size(&self) -> usize {
-        const ELEM_SIZE: usize = core::mem::size_of::<BitSliceType>();
-
-        if self.is_inline() {
-            return 0;
-        }
-
-        let len = unsafe { self.len_unchecked() } + 1;
-        ELEM_SIZE * len
-    }
-}
-
 #[cfg(test)]
 mod tests {
     #![allow(clippy::unwrap_used)]
 
-    use crate::bst_slice::BstSlice;
-
     use super::*;
 
     #[cfg(not(feature = "std"))]
-    use extern_alloc::{
-        string::{String, ToString},
-        vec::Vec,
-    };
-
-    /// Helper extension trait to simplify some tests.
-    trait AndNot {
-        fn and_not(self, rhs: Self) -> Self;
-    }
-
-    impl AndNot for BitSliceType {
-        fn and_not(self, rhs: Self) -> Self {
-            self & !rhs
-        }
-    }
-
-    impl AndNot for usize {
-        fn and_not(self, rhs: Self) -> Self {
-            self & !rhs
-        }
-    }
-
-    fn to_bst_vec(sbs: &SmolBitSet) -> Vec<BitSliceType> {
-        BstSlice::new(sbs).slice().to_vec()
-    }
-
-    fn zero_pad_to(mut vec: Vec<BitSliceType>, len: usize) -> Vec<BitSliceType> {
-        while vec.len() < len {
-            vec.push(0);
-        }
-        vec
-    }
+    use extern_alloc::string::{String, ToString};
 
     #[test]
     fn send() {
@@ -738,80 +678,6 @@ mod tests {
         assert_eq!(sbs.to_string(), "220179738009501684669546686565819917733");
     }
 
-    #[test]
-    #[cfg(feature = "std")]
-    fn hash() {
-        // core does not have a default hasher
-        use hash::{DefaultHasher, Hash, Hasher};
-
-        let a = SmolBitSet::from(0xC5C5_F00Du32);
-        let b = SmolBitSet::from(0xC5C5_F00Du32);
-        let mut hasher_a = DefaultHasher::new();
-        let mut hasher_b = DefaultHasher::new();
-        a.hash(&mut hasher_a);
-        b.hash(&mut hasher_b);
-        assert_eq!(hasher_a.finish(), hasher_b.finish());
-
-        let mut a = SmolBitSet::from(0xFFC5_C0FF_EE00_BEEF_u64);
-        let mut b = SmolBitSet::from(0xFFC5_C0FF_EE00_BEEF_u64);
-        a <<= 128u8;
-        a >>= 66u8;
-        b <<= 128u8 - 66u8;
-        let mut hasher_a = DefaultHasher::new();
-        let mut hasher_b = DefaultHasher::new();
-        a.hash(&mut hasher_a);
-        b.hash(&mut hasher_b);
-        assert_eq!(hasher_a.finish(), hasher_b.finish());
-    }
-
-    mod from {
-        use super::*;
-
-        macro_rules! test_from {
-            ($($name:ident, $val:expr),+) => {$(
-                #[test]
-                fn $name() {
-                    let t = SmolBitSet::from($val);
-                    assert!(t.is_inline());
-
-                    let d = unsafe { t.get_inline_data_unchecked() };
-                    assert_eq!(d, $val as usize);
-                }
-            )*}
-        }
-
-        test_from! {
-            u8, 0b1110_1010u8,
-            u8_max, u8::MAX,
-            u16, 0xBEAFu16,
-            u16_max, u16::MAX,
-            u32, 0xF0BA_B0BBu32,
-            u32_max, u32::MAX,
-            u64, 0x3550_1337_0000_F00Fu64
-        }
-
-        #[test]
-        fn u64_hb_64() {
-            let t = SmolBitSet::from(0xC5C5_BEEF_0000_1234u64);
-            assert!(!t.is_inline());
-            assert_eq!(t.len(), 2);
-
-            let d = t.as_slice();
-            assert_eq!(d.len(), 2);
-            assert_eq!(d, [0x0000_1234, 0xC5C5_BEEF]);
-        }
-
-        #[test]
-        fn u64_max() {
-            let t = SmolBitSet::from(u64::MAX);
-            assert_eq!(t.len(), 2);
-
-            let d = t.as_slice();
-            assert_eq!(d.len(), 2);
-            assert_eq!(d, &[u32::MAX; 2]);
-        }
-    }
-
     mod clone {
         use super::*;
 
@@ -845,409 +711,6 @@ mod tests {
             assert_eq!(a_data.len(), b_data.len());
             assert_eq!(a_data, b_data);
             assert_eq!(a_data, [0x1337_0420, 0xFFEE_00AA]);
-        }
-    }
-
-    mod shifts {
-        use super::*;
-
-        mod shl {
-            use super::*;
-
-            #[test]
-            fn inline() {
-                let val = 0xABCD_0042u32;
-                let mut a = SmolBitSet::from(val);
-                assert!(a.is_inline());
-
-                a <<= 6u8;
-                assert!(a.is_inline());
-                assert_eq!(
-                    unsafe { a.get_inline_data_unchecked() },
-                    (val as usize) << 6
-                );
-
-                let b = a << 4u8;
-                assert!(b.is_inline());
-                assert_eq!(
-                    unsafe { b.get_inline_data_unchecked() },
-                    (val as usize) << (6 + 4)
-                );
-            }
-
-            #[test]
-            fn grow_to_slice() {
-                let val = 0x00AB_CD55_1337_BEEFu64;
-                let mut a = SmolBitSet::from(val);
-                assert!(a.is_inline());
-
-                a <<= 8u8;
-                assert_eq!(a.len(), 2);
-                assert_eq!(a.as_slice(), [0x37BE_EF00u32, 0xABCD_5513u32]);
-
-                let b = a << 24u8;
-                assert_eq!(b.len(), 3);
-                assert_eq!(
-                    b.as_slice(),
-                    [0x0000_0000u32, 0x1337_BEEFu32, 0x00AB_CD55u32]
-                );
-            }
-
-            #[test]
-            fn by_multiple_of_32() {
-                let val = 0xFFEE_00AA_AFFE_BEEFu64;
-                let mut a = SmolBitSet::from(val);
-                assert!(!a.is_inline());
-
-                a <<= 32u8;
-                assert_eq!(a.len(), 3);
-                assert_eq!(a.as_slice(), [0, 0xAFFE_BEEFu32, 0xFFEE_00AAu32]);
-
-                a <<= 64u8;
-                assert_eq!(a.len(), 5);
-                assert_eq!(a.as_slice(), [0, 0, 0, 0xAFFE_BEEFu32, 0xFFEE_00AAu32]);
-            }
-        }
-
-        mod shr {
-            use super::*;
-
-            #[test]
-            fn inline() {
-                let val = 0xF00D_BEEFu32;
-                let mut a = SmolBitSet::from(val);
-                assert!(a.is_inline());
-
-                a >>= 10u8;
-                assert!(a.is_inline());
-                assert_eq!(
-                    unsafe { a.get_inline_data_unchecked() },
-                    (val >> 10) as usize
-                );
-
-                let b = a >> 10u8;
-                assert!(b.is_inline());
-                assert_eq!(
-                    unsafe { b.get_inline_data_unchecked() },
-                    (val >> (10 + 10)) as usize
-                );
-            }
-
-            #[test]
-            fn slice() {
-                let val = 0xF420_1337_FEFE_BEEFu64;
-                let mut a = SmolBitSet::from(val);
-                assert!(!a.is_inline());
-                assert_eq!(a.len(), 2);
-
-                a >>= 8u8;
-                assert!(!a.is_inline());
-                assert_eq!(a.as_slice(), [0x37FE_FEBEu32, 0x00F4_2013u32]);
-
-                let b = a >> 24u8;
-                assert!(!b.is_inline());
-                assert_eq!(b.as_slice(), [0xF420_1337u32, 0x0000_0000u32]);
-            }
-
-            #[test]
-            fn by_multiple_of_32() {
-                let val = 0xFFEE_00AA_AFFE_BEEFu64;
-                let mut a = SmolBitSet::from(val);
-                assert!(!a.is_inline());
-
-                a >>= 32u8;
-                assert_eq!(a.len(), 2);
-                assert_eq!(a.as_slice(), [0xFFEE_00AAu32, 0]);
-
-                let mut a = SmolBitSet::from(val);
-                a >>= 64u8;
-                assert_eq!(a.len(), 2);
-                assert_eq!(a.as_slice(), [0, 0]);
-            }
-        }
-
-        #[test]
-        fn with_offsets() {
-            let val = 0xA5A5_BEEF_1337_A5A5u64;
-            let mut a = SmolBitSet::from(val);
-            assert!(!a.is_inline());
-            assert_eq!(a.len(), 2);
-
-            a <<= 64u16 + 24u16;
-            assert!(!a.is_inline());
-            assert_eq!(a.len(), 2 + 2 + 1);
-            assert_eq!(a.as_slice(), [0, 0, 0xA500_0000, 0xEF13_37A5, 0x00A5_A5BE]);
-
-            a >>= 32u32 + 16u32;
-            assert!(!a.is_inline());
-            assert_eq!(a.len(), 2 + 2 + 1); // still same size, does not auto shrink
-            assert_eq!(a.as_slice(), [0, 0x37A5_A500, 0xA5BE_EF13, 0x0000_00A5, 0],);
-        }
-    }
-
-    mod binops {
-        use super::*;
-
-        use core::ops::{BitAnd, BitOr, BitXor};
-
-        mod inline {
-            use super::*;
-
-            macro_rules! test_inline_binops {
-                ($($name:ident, $a:expr, $b:expr),+) => {$(
-                    #[test]
-                    fn $name() {
-                        let a = SmolBitSet::from($a);
-                        let b = SmolBitSet::from($b);
-                        assert!(a.is_inline());
-                        assert!(b.is_inline());
-
-                        let res = a.$name(&b);
-                        assert!(res.is_inline());
-
-                        let d = unsafe { res.get_inline_data_unchecked() };
-                        assert_eq!(d, ($a as usize).$name($b as usize));
-                    }
-                )*}
-            }
-
-            test_inline_binops! {
-                bitor, 0xF0F0u16, 0x0F0Fu16,
-                bitand, 0xFEFEu16, 0xAFFEu16,
-                bitxor, 0x3A52u16, 0xAAE8u16,
-                and_not, 0xFEFEu16, 0xAFFEu16
-            }
-        }
-
-        mod slice {
-            use super::*;
-
-            macro_rules! test_slice_binops {
-                ($($name:ident, $a:expr, $b:expr),+) => {$(
-                    #[test]
-                    fn $name() {
-                        let a = SmolBitSet::from($a);
-                        let b = SmolBitSet::from($b);
-                        assert_eq!(a.len(), 2);
-                        assert_eq!(b.len(), 2);
-
-                        let res = a.$name(&b);
-                        assert_eq!(res.len(), 2);
-                        assert_eq!(
-                            res.as_slice(),
-                            [
-                                (($a as BitSliceType).$name($b as BitSliceType)),
-                                ((($a >> 32) as BitSliceType).$name(($b >> 32) as BitSliceType))
-                            ]
-                        );
-                    }
-                )*}
-            }
-
-            test_slice_binops! {
-                bitor, 0xC0FF_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                bitand, 0xC0FF_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                bitxor, 0xC0FF_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                and_not, 0xC0FF_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64
-            }
-        }
-
-        mod mixed {
-            use super::*;
-
-            macro_rules! test_mixed_binops {
-                ($($name:ident, $a:expr, $b:expr),+) => {$(
-                    #[test]
-                    fn $name() {
-                        let a = SmolBitSet::from($a);
-                        let b = SmolBitSet::from($b);
-                        assert!(a.is_inline());
-                        assert_eq!(b.len(), 2);
-
-                        let res1 = a.clone().$name(&b);
-                        assert_eq!(
-                            to_bst_vec(&res1),
-                            [
-                                (($a as BitSliceType).$name($b as BitSliceType)),
-                                ((($a >> 32) as BitSliceType).$name(($b >> 32) as BitSliceType))
-                            ]
-                        );
-
-                        let res2 = b.$name(&a);
-                        assert_eq!(res2, res1);
-                    }
-                )*}
-            }
-
-            test_mixed_binops! {
-                bitor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                bitand, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                bitxor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64
-            }
-
-            // `a.and_not(b) != b.and_not(a)`
-            #[test]
-            fn and_not() {
-                const A: u64 = 0x0ABC_EE00_1337_BEEFu64;
-                const B: u64 = 0xF00D_BEEF_0420_BEEFu64;
-
-                let a = SmolBitSet::from(A);
-                let b = SmolBitSet::from(B);
-                assert!(a.is_inline());
-                assert_eq!(b.len(), 2);
-
-                let res1 = a.and_not(&b);
-                assert_eq!(
-                    to_bst_vec(&res1),
-                    [
-                        ((A as BitSliceType).and_not(B as BitSliceType)),
-                        (((A >> 32) as BitSliceType).and_not((B >> 32) as BitSliceType))
-                    ]
-                );
-            }
-        }
-
-        mod rhs_smaller {
-            use super::*;
-
-            macro_rules! test_rhs_smaller_binops {
-                ($($name:ident, $a:expr, $b:expr),+) => {$(
-                    #[test]
-                    fn $name() {
-                        let mut lhs = SmolBitSet::from($a);
-                        let rhs = SmolBitSet::from($b);
-                        lhs <<= 32u8;
-                        let lhs_len = lhs.len();
-                        assert!(lhs_len > rhs.len());
-
-                        let res = lhs.$name(&rhs);
-                        assert_eq!(
-                            to_bst_vec(&res),
-                            [
-                                ((0 as BitSliceType).$name($b as BitSliceType)),
-                                ((($a) as BitSliceType).$name(($b >> 32) as BitSliceType)),
-                                ((($a >> 32) as BitSliceType).$name(0 as BitSliceType))
-                            ]
-                        );
-                    }
-                )*};
-            }
-
-            test_rhs_smaller_binops! {
-                bitor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                bitand, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                bitxor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                and_not, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64
-            }
-        }
-
-        mod rhs_larger {
-            use super::*;
-
-            macro_rules! test_rhs_larger_binops {
-                ($($name:ident, $a:expr, $b:expr),+) => {$(
-                    #[test]
-                    fn $name() {
-                        let lhs = SmolBitSet::from($a);
-                        let mut rhs = SmolBitSet::from($b);
-                        rhs <<= 32u8;
-                        let rhs_len = rhs.len();
-                        assert!(rhs_len > lhs.len());
-
-                        let res = lhs.$name(&rhs);
-                        assert_eq!(
-                            zero_pad_to(to_bst_vec(&res), 3),
-                            [
-                                (($a as BitSliceType).$name(0 as BitSliceType)),
-                                ((($a >> 32) as BitSliceType).$name($b as BitSliceType)),
-                                ((0 as BitSliceType).$name(($b >> 32) as BitSliceType))
-                            ]
-                        );
-                    }
-                )*};
-            }
-
-            test_rhs_larger_binops! {
-                bitor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                bitand, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                bitxor, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64,
-                and_not, 0x0ABC_EE00_1337_BEEFu64, 0xF00D_BEEF_0420_BEEFu64
-            }
-        }
-    }
-
-    mod cmp {
-        use super::*;
-
-        #[test]
-        fn eq() {
-            let mut a = SmolBitSet::from(u16::MAX);
-            let mut b = SmolBitSet::from(0xFFFFu16);
-            assert_eq!(a, b);
-
-            a <<= 55;
-            assert_ne!(a, b);
-
-            b <<= 55;
-            assert_eq!(a, b);
-        }
-
-        #[test]
-        fn ord() {
-            let mut a = SmolBitSet::from(0xBEEFu16);
-            let mut b = SmolBitSet::from(0x00C5_F00Du32);
-            assert!(a < b);
-
-            a <<= 72;
-            assert!(a > b);
-            assert!(b < a);
-
-            b <<= 72;
-            assert!(a < b);
-        }
-
-        #[test]
-        fn eq_but_not_physical_same() {
-            let mut a = SmolBitSet::from(u16::MAX);
-            let mut b = SmolBitSet::from(0xFFFFu16);
-
-            // ensure a is larger than b in memory
-            a.spill(256);
-
-            assert_eq!(a, b);
-            assert_eq!(b, a);
-
-            a <<= 55;
-            assert_ne!(a, b);
-            assert_ne!(b, a);
-
-            b <<= 55;
-            assert_eq!(a, b);
-            assert_eq!(b, a);
-        }
-
-        #[test]
-        fn ord_but_not_physical_same() {
-            let mut a = SmolBitSet::from(0xBEEFu16);
-            let mut b = SmolBitSet::from(0x00C5_F00Du32);
-
-            // ensure a is larger than b in memory
-            a.spill(256);
-
-            assert!(a < b);
-            assert!(b > a);
-
-            a <<= 18;
-            assert!(a > b);
-            assert!(b < a);
-
-            a <<= 54;
-            assert!(a > b);
-            assert!(b < a);
-
-            b <<= 72;
-            assert!(a < b);
-            assert!(b > a);
         }
     }
 }
